pyproject.toml
[tool.poetry]
name = "pusher-test"
version = "0.1.0"
description = ""
authors = ["Somnath Banerjee <somnath.banerjee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.4"
pusher = "^3.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
